Removing Transfer-Encoding and switching to Content-Length
