+++ core/amazonka-core.cabal
@@ -65,6 +65,10 @@ library
         , Network.AWS.Types
         , Network.AWS.Waiter
 
+    other-modules:
+          Network.AWS.Sign.V4.Base
+        , Network.AWS.Sign.V4.Chunked
+
     build-depends:
           aeson                >= 0.8
         , attoparsec           >= 0.11.3
